fixed: corrected permissions for SSL private key
The SSL private key is now set to not be world readable. 

In addition the `combined.pem` (aka `fullchain.pem`) was incorrectly including the private key. The `fullchain.pem` files should never include the private key. They contain only the certificate and the CA certificates necessary to validate the certificate. In most cases the CA certificates would be user supplied for CA(s) that are not included in the `ca-certificates` package. 

Signed-off-by: Gerard Hickey <hickey@kinetic-compute.com>